[#11123] geo: Safe deletion of transaction tables
Summary:
This diff makes the DeleteTable RPC (and by extension, `yb-admin delete_table`) work
safely for transaction tables.

Deletion of transaction tables is carried out in the following manner:

1. The table is marked as DELETING, and removed from the set of transaction tables used to pick
   status tablets for new transactions.
2. The transaction tables version is incremented, and propagated to all tservers. At this point,
   no newly created transactions will use any tablet from the transaction table being deleted.
3. A PrepareDeleteTransactionTablet RPC is sent to each status tablets' leader. This RPC waits
   for all running transactions to either abort, or commit and apply in all involved tablets.
4. DeleteTablet RPCs are sent out as usual to each replica of each tablet.

Step 3 ensures that all committed transactions' intents have been applied successfully, but does
not ensure that all aborted transactions' intents have been cleaned up. This diff modifies the
behavior of the transaction status resolver to assume all transactions on a transaction tablet
have been aborted, if the transaction tablet has been deleted.

Transaction tables are also now automatically deleted when a tablespace is dropped.

Test Plan: `ybd --gtest_filter GeoTransactionsTest.TestTransactionTableDeletion`

Reviewers: rsami, sergei, asrivastava

Reviewed By: sergei

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D17791

+++ src/yb/tserver/ts_tablet_manager.cc
@@ -1273,6 +1273,7 @@ Status TSTabletManager::DeleteTablet(
     tablet::ShouldAbortActiveTransactions should_abort_active_txns,
     const boost::optional<int64_t>& cas_config_opid_index_less_or_equal,
     bool hide_only,
+    bool keep_data,
     boost::optional<TabletServerErrorPB::Code>* error_code) {
 
   if (delete_type != TABLET_DATA_DELETED && delete_type != TABLET_DATA_TOMBSTONED) {
@@ -1347,20 +1348,22 @@ Status TSTabletManager::DeleteTablet(
 
   yb::OpId last_logged_opid = tablet_peer->GetLatestLogEntryOpId();
 
-  Status s = DeleteTabletData(meta,
-                              delete_type,
-                              fs_manager_->uuid(),
-                              last_logged_opid,
-                              this);
-  if (PREDICT_FALSE(!s.ok())) {
-    s = s.CloneAndPrepend(Substitute("Unable to delete on-disk data from tablet $0",
-                                     tablet_id));
-    LOG(WARNING) << s.ToString();
-    tablet_peer->SetFailed(s);
-    return s;
-  }
+  if (!keep_data) {
+    Status s = DeleteTabletData(meta,
+                                delete_type,
+                                fs_manager_->uuid(),
+                                last_logged_opid,
+                                this);
+    if (PREDICT_FALSE(!s.ok())) {
+      s = s.CloneAndPrepend(Substitute("Unable to delete on-disk data from tablet $0",
+                                       tablet_id));
+      LOG(WARNING) << s.ToString();
+      tablet_peer->SetFailed(s);
+      return s;
+    }
 
-  tablet_peer->status_listener()->StatusMessage("Deleted tablet blocks from disk");
+    tablet_peer->status_listener()->StatusMessage("Deleted tablet blocks from disk");
+  }
 
   // We only remove DELETED tablets from the tablet map.
   if (delete_type == TABLET_DATA_DELETED) {
